requirements.txt
py-clob-client>=0.20.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
web3>=6.0.0
python-telegram-bot>=20.0
colorlog>=6.7.0
